Crate cfx_rpc_eth_types

Re-exports

pub use trace_filter::TraceFilter;

pub use eth_pubsub::*;

pub use trace::*;

Modules

eth_pubsub

Pub-Sub types.

trace

trace_filter

Trace filter deserialization.

Structs

AccessListResult

AccessListResult for handling errors from eth_createAccessList

AccessListWithGasUsed

Access list with gas used appended.

AccountOverride

AccountPendingTransactions

Authorization

Block

Block representation

BlockOverrides

BlockOverrides is a set of header fields to override.

Bundle

Bundle of transactions

Bytes

Wrapper structure around vector of bytes.

EthCallResponse

CallResponse for eth_callMany

EthRpcLogFilter

Filter

EvmOverrides

Helper type that bundles various overrides for EVM Execution.

FeeHistory

Header

Block header representation.

Index

Represents usize.

Log

Log

Receipt

Receipt

RpcAccountOverride

Custom account override used in rpc call

SignedAuthorization

SimBlock

Represents a batch of calls to be simulated sequentially within a block. This struct includes block and state overrides as well as the transaction requests to be executed.

SimCallResult

Captures the outcome of a transaction simulation. It includes the return value, logs produced, gas used, and the status of the transaction.

SimulateError

The error response returned by the eth_simulateV1 method.

SimulatePayload

Simulation options for executing multiple blocks and transactions.

SimulatedBlock

Represents the result of simulating a block.

StateContext

State context for callMany

SyncInfo

Sync info

Transaction

Transaction

TransactionRequest

Call request

TxpoolContent

TxpoolContentFrom

Transaction Pool Content From

TxpoolInspect

Transaction Pool Inspect

TxpoolInspectSummary

Transaction summary as found in the Txpool Inspection property.

TxpoolStatus

U64

Enums

AccountStateOverrideMode

BlockNumber

Represents rpc api block number param.

Error

FilterChanges

Results of the filter_changes RPC.

SyncStatus

Sync status

TransactionIndex

Represents a transaction index where -1 means all transactions

Constants

DEFAULT_ETH_GAS_CALL_REQUEST

MAX_SIMULATE_BLOCKS

The maximum number of blocks that can be simulated in a single request,

Type Aliases

FilterAddress

Filter Address

RpcStateOverride

A set of account overrides

StateOverride

Topic

Topic